+++ pyage/audio_backends/sound_wrapper.py
@@ -22,4 +22,13 @@ class SoundWrapper(ABC):
 
     @volume.setter
     def volume(self, value: float) -> None:
+        raise NotImplementedError
+
+    @property
+    @abstractmethod
+    def looping(self) -> bool:
         pass
+
+    @looping.setter
+    def looping(self, looping: bool) -> None:
+        raise NotImplementedError